Autosteer button active now, although there's not yet any means to actually transmit data to steering unit.
Tried to make the top line display look better, but the layout is still
not quite right.  the txtDistanceABLine field is way too big and crowds
other fields.  Not sure why.  Eventually I want to change these fields
to QML and place them at the top of the QML view (moving the sky and
lightbar down).