CModuleComm updated to AOG bf871446b08f86579c197822463810a8011970b5
Unfortunately we can't refer to main form in this class due to circular
references that we want to avoid, so the caller will have to call
rateRelayControlOutToPort and autoSteerDataOutToPort on the main form
after calling the resetAllModuleCommValues() method.

Maybe in the future we could Qt events to allow resetAllModuleCommValues
to trigger calls to those functions.